python/requirements.txt
black
pandas
python-dotenv
requests
